Refactoring: extract TFileRingBuffer, TDynamicPersistentTable, TPersistentTable to a separate library
