+++ cloud/filestore/libs/vfs_fuse/fuzz/starter.cpp
@@ -16,10 +16,10 @@
 #include <cloud/filestore/libs/vhost/server.h>
 
 #include <cloud/storage/core/libs/common/error.h>
-#include <cloud/storage/core/libs/common/file_map_memory_limiter.h>
 #include <cloud/storage/core/libs/common/scheduler.h>
 #include <cloud/storage/core/libs/common/timer.h>
 #include <cloud/storage/core/libs/diagnostics/logging.h>
+#include <cloud/storage/core/libs/file_backed_containers/file_map_memory_limiter.h>
 #include <cloud/storage/core/libs/vhost-client/vhost-buffered-client.h>
 
 #include <cloud/contrib/vhost/virtio/virtio_fs_spec.h>
