+++ cloud/filestore/libs/vfs_fuse/write_back_cache/ya.make
@@ -30,6 +30,7 @@ PEERDIR(
 
     cloud/storage/core/libs/common
     cloud/storage/core/libs/diagnostics
+    cloud/storage/core/libs/file_backed_containers
 
     library/cpp/threading/future
     library/cpp/threading/future/subscription
