+++ cloud/storage/core/libs/file_backed_containers/file_ring_buffer.cpp
@@ -789,7 +789,7 @@ public:
 
     void SetCorrupted()
     {
-        Corrupted = true;
+            Corrupted = true;
     }
 
     ui64 GetRawCapacity() const
