+++ cloud/storage/core/libs/file_backed_containers/file_ring_buffer.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "error.h"
+#include <cloud/storage/core/libs/common/error.h>
 
 #include <util/generic/string.h>
 #include <util/generic/vector.h>
